Reorganize in a hopefully sensible way.
